payment

+++ engine/src/abs/Client.java
@@ -77,7 +77,7 @@ public class Client implements Serializable {
         asGiver.add(loan);
     }
 
-    public void subtractCurrBalance(int amountToDeducted) {
+    public void subtractCurrBalance(double amountToDeducted) {
         currBalance -= amountToDeducted;
     }
     public void addToCurrBalance(double amountToAdd) {
